// src/app/modules/director/director.interface.ts
export type TCompanyData = {
  ucin: string;
  din: string;
  cin_LLPIN: string;
  nameOfTheCompany: string;
  emailAddress: string;
  role: string;
  roleEffectiveDate: string;
  cessationDate: string;
  directorFlag: string;
  designation: string;
  accountType: string;
  dinAllocationDate: string;
  bodyCorporateType: string;
  profitSharingPercentage: string;
  drivingLicenseNumber: string;
  occupationType: string;
  citizenOfIndia: string;
  bodyCorpOutsideIndiaIDNum: string;
  authorizationStatus: string;
  companyStatus: string;
  bodyCorpInsideIndiaIDNum: string;
  recordCreatedDate: string;
  userName: string;
  companyType: string;
  userId: string;
  dinSurrenderDate: string;
  roleLICValue: string;
  bodyCorpOutsideIndiaId: string;
  middleName: string;
  firstName: string;
  bodyCorpOutsideIndiaName: string;
  companyClass: string;
  residentOfIndia: string;
  fathersLastName: string;
  opcType: string;
  monetaryContributionValue: string;
  companyOrigin: string;
  v2UserId: string;
  pan: string;
  fathersMiddleName: string;
  migrationFlag: string;
  passportNumber: string;
  approverId: string;
  personType: string;
  flagged: string;
  mobileNumber: string;
  fllpDesignation: string;
  companyId: string;
  fathersFirstName: string;
  directorCategory: string;
  nationality: string;
  directorDeathDate: string;
  oldFlag: string;
  srn: string;
  dob: string;
  membershipNumber: string;
  contributionForm: string;
  sboFlag: string;
  bodyCorpInsideIndiaId: string;
  migrationDate: string;
  educationalQualification: string;
  otherOccupation: string;
  compAccountType: string;
  bodyCorpInsideIndiaName: string;
  occupation: string;
  signatoryAssociationStatus: string;
  gender: string;
  isDisqualified: string;
  obligationContribution: string;
  recordUpdatedDate: string;
  shareholdingPercentage: string;
  birthPlace: string;
  opcFlag: string;
  othersEducationalQualification: string;
  lastName: string;
  currentDesignationDate: string;
};